Add update via handler index for handlers
This allows us to release new handlers that will be deployed within a day of running `bundle install`
It works by downloading the `index` from master on github pages and checking existence and mtimes
of all files we have locally. If we are missing or outdated on a file, we download from github pages

This runs at most once per day and slows down the command by about 1 second that one time. Otherwise there is almost no impact.